Add prettier and prettier-solidity.

+++ contracts/Migrations.sol
@@ -3,23 +3,23 @@
 pragma solidity >=0.6.6 <0.7.0;
 
 contract Migrations {
-  address public owner;
-  uint public lastCompletedMigration;
+    address public owner;
+    uint256 public lastCompletedMigration;
 
-  constructor() public {
-    owner = msg.sender;
-  }
+    constructor() public {
+        owner = msg.sender;
+    }
 
-  modifier restricted() {
-    if (msg.sender == owner) _;
-  }
+    modifier restricted() {
+        if (msg.sender == owner) _;
+    }
 
-  function setCompleted(uint completed) public restricted {
-    lastCompletedMigration = completed;
-  }
+    function setCompleted(uint256 completed) public restricted {
+        lastCompletedMigration = completed;
+    }
 
-  function upgrade(address newAddress) public restricted {
-    Migrations upgraded = Migrations(newAddress);
-    upgraded.setCompleted(lastCompletedMigration);
-  }
+    function upgrade(address newAddress) public restricted {
+        Migrations upgraded = Migrations(newAddress);
+        upgraded.setCompleted(lastCompletedMigration);
+    }
 }
